chore: evict Curve2D.swift's foreign material into dedicated files (#687)
Curve2D.swift carried 21 declarations belonging to types other than Curve2D
(the Gcc constraint solver family, analytic intersection/extrema namespaces).
Evicts them, following the #393/#394 "evict before you split" pattern.

Moved to new Curve2DGcc.swift (720 lines):
  Curve2DGcc (enum + both extensions), Curve2DQualifier, Curve2DCircleSolution,
  Curve2DLineSolution, Curve2DHatchSegment, BisecType, BisecSolution, GccAnaBisector

Moved to new IntAna2d.swift (95 lines):
  Intersection2DPoint, IntAna2d

Moved to new Extrema2d.swift (156 lines):
  Extrema2DResult, Extrema2d

Kept in Curve2D.swift: Curve2DIntersection, Curve2DProjection, Curve2DExtremaResult,
Curve2DSpecialPointType, Curve2DSpecialPoint, CurInfType (+ extension), CurInfPoint.
These are result types returned exclusively by Curve2D's own methods, the same
role Curve3D.swift's own CurveExtremaResult/CurveSurfaceHit already play there
(the sibling file's accepted 0.3% foreign material) rather than a separate
algorithm family, so they stay (60/3362 lines, 1.8%, same order of magnitude).

Also moves Surface.swift's CurveSurfaceIntersection + extension Curve3D (both
genuinely Curve3D's own, unlike Surface's own SurfaceGrid/SurfaceGridD1/
BezierFillStyle which stay) onto Curve3D.swift.

Curve2D.swift: 4324 -> 3362 lines. Surface.swift: 5388 -> 5353 lines.
Curve3D.swift: 4121 -> 4156 lines. Zero behavior change: bodies moved verbatim,
same module, same public API.
